Struct mentat_core::StructuredMap

pub struct StructuredMap(pub IndexMap<ValueRc<Keyword>, Binding>);

A pull expression expands a binding into a structure. The returned structure associates attributes named in the input or retrieved from the store with values. This association is a StructuredMap.

Note that 'attributes' in Datomic's case can mean:

Reversed attribute keywords (:artist/_country).

An alias using :as (:artist/name :as "Band name").

We entirely support the former, and partially support the latter -- you can alias using a different keyword only.

Methods

impl StructuredMap

pub fn insert<N, B>(&mut self, name: N, value: B) where
N: Into<ValueRc<Keyword>>,
B: Into<Binding>,